tests(project): fixed test setup
