+++ tests/unit-tests/UnitTests/ContentFactory/SuccessPageContentFactoryTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using form_builder.Builders;
+using form_builder.Configuration;
 using form_builder.ContentFactory.PageFactory;
 using form_builder.ContentFactory.SuccessPageFactory;
 using form_builder.Enum;
@@ -15,6 +16,7 @@ using form_builder.Services.PageService.Entities;
 using form_builder.ViewModels;
 using form_builder_tests.Builders;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Options;
 using Moq;
 using Xunit;
 
@@ -28,14 +30,20 @@ namespace form_builder_tests.UnitTests.ContentFactory
         private readonly Mock<ISessionHelper> _mockSessionHelper = new ();
         private readonly Mock<IDistributedCacheWrapper> _mockDistributedCache = new ();
         private readonly Mock<IWebHostEnvironment> _mockWebHostEnvironment = new ();
+        private readonly Mock<IOptions<PreviewModeConfiguration>> _mockPreviewModeConfiguration = new ();
 
         public SuccessPageContentFactoryTests()
         {
+            _mockPreviewModeConfiguration
+                .Setup(_ => _.Value)
+                .Returns(new PreviewModeConfiguration());
+
             _factory = new SuccessPageFactory(
                 _mockPageHelper.Object,
                 _mockPageContentFactory.Object,
                 _mockSessionHelper.Object,
                 _mockDistributedCache.Object,
+                _mockPreviewModeConfiguration.Object,
                 _mockWebHostEnvironment.Object);
 
             _mockWebHostEnvironment.Setup(_ => _.EnvironmentName).Returns("local");
